Fixed feedback handler states filters.

+++ msk_zoo_tg_bot/admin_and_models/management/commands/zoo_in_telega/handlers/random_messages_handler.py
@@ -5,6 +5,7 @@ from aiogram import types, Dispatcher
 from aiogram.dispatcher import FSMContext
 
 from filters.random_messages_filter import random_message_filter
+from handlers.admin_panel_handlers import admin_delete_spam_handler
 
 
 # ------------------------
@@ -12,12 +13,19 @@ from filters.random_messages_filter import random_message_filter
 async def random_message_handler(message: types.Message, state: FSMContext) -> None:
     """Функция для удаления рандомных сообщений от пользователя."""
 
-    cur_state = state.get_state()
+    cur_state = await state.get_state()
 
-    await message.delete()
-    logging.info(f' {datetime.now()} : User with ID = {message.from_user.id} and username = '
-                 f'{message.from_user.username} tried to send random message '
-                 f'with {message.content_type} type while in {cur_state} state.')
+    if cur_state == 'Feedback:feedback':
+        return
+
+    elif cur_state == 'AdminAuthorization:TRUE':
+        await admin_delete_spam_handler(message=message)
+
+    else:
+        await message.delete()
+        logging.info(f' {datetime.now()} : User with ID = {message.from_user.id} and username = '
+                     f'{message.from_user.username} tried to send random message '
+                     f'with {message.content_type} type while in {cur_state} state.')
 
 
 # ---------------------
